Fix up the ACLs database.  Split out non-statics into a .h file.
